ENH: Use Resample Scalar/Vector/DWI Volume module in SegmentAxisAlignment
After some experiments it turned out that the Crop Volume module does excessive interpolation (smoothing) on the volume when rotating it. Trying the available resampling volumes it seems that the Resample Scalar/Vector/DWI Volume yields the clearest result.

Remove the supersample/subsample steps from the module because it does not improve the quality of the output image.